pyproject.toml
[tool.poetry]
name = "itmo-mts-recsys-2023"
version = "0.0.1"
description = ""
authors = ["Emiliy Feldman <feldlime@yandex.ru>"]
maintainers = ["Emiliy Feldman <feldlime@yandex.ru>"]
readme = "README.md"
packages = [
    { include = "service" }
]

[tool.poetry.dependencies]
python = "^3.10.12"
fastapi = "^0.104.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
uvloop = "^0.19.0"
gunicorn = "^21.2.0"
orjson = "^3.9.10"
httpx = "^0.25.1"
uvicorn = "^0.24.0.post1"
pyjwt = "^2.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pylint = "^3.0.2"
black = "^23.11.0"

[tool.black]
line-length = 120
target-version = ["py310"]

[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"
